+++ src/stores/types/httpTypes.ts
